Updated agreetos

+++ main.py
@@ -182,7 +182,7 @@ async def on_message(message):
 	timestamp = str(time.time())
 	userid = message.author.id
 
-	if message.content.startswith("/") and message.content != "/agreetos" and message.content != "/cagreedtos" and message.content != "/help" and userid in agreetos and userid in ragreedtos or message.author.id == "409090118956089344":
+	if message.content.startswith("/") and message.content != "/agreetos" and message.content != "/ragreedtos" and message.content != "/cagreedtos" and message.content != "/help" and userid in agreetos and userid in ragreedtos or message.author.id == "409090118956089344":
 		# 全件取得は cursor.fetchall()
 		# 「/register」で始まるか調べる
 		if message.content.startswith("/"):
@@ -1166,7 +1166,6 @@ async def on_message(message):
 
 
 	if message.content.startswith("/"):
-		#共用コマンド
 		if message.content == "/cagreedtos":
 			start = time.time()
 				# データベース接続とカーソル生成
@@ -1175,7 +1174,7 @@ async def on_message(message):
 			await client.add_reaction(message, '👌')
 			fee = "0.01"
 			cursor.execute("INSERT INTO agreetos (id) VALUES (%s)", (username,))
-			m = "<@" + userid + "> 利用規約への同意を確認しました。"
+			m = "<@" + userid + "> 利用規約への同意を確認しました。Monageへの協力ありがとうございました。おみくじでもらえる金額を減らしておきますね！"
 			await client.send_message(message.channel, m)
 			cursor.execute('SELECT * FROM agreetos')
 			agreetos = cursor.fetchall()
@@ -1187,19 +1186,24 @@ async def on_message(message):
 			# 保存を実行（忘れると保存されないので注意）
 			connection.commit()
 		if message.content == "/ragreedtos":
-			start = time.time()
-				# データベース接続とカーソル生成
-			# エラー処理（例外処理）
-			await client.add_reaction(message, '👌')
-			fee = "0.01"
-			cursor.execute("INSERT INTO ragreedtos (id) VALUES (%s)", (userid,))
-			m = "<@" + userid + "> 利用規約への同意を確認しました。"
-			await client.send_message(message.channel, m)
-			cursor.execute('SELECT * FROM ragreedtos')
-			agreetos = cursor.fetchall()
-			agreetos = str(agreetos)
-			print(agreetos)
-			await client.delete_message(message)
+			cursor.execute('SELECT * FROM agreetos')
+			agreedtos = cursor.fetchall()
+			agreedtos = mlibs.fixselect(agreetos)
+			#共用コマンド
+			if userid not in agreedtos:
+				start = time.time()
+					# データベース接続とカーソル生成
+				# エラー処理（例外処理）
+				await client.add_reaction(message, '👌')
+				fee = "0.01"
+				cursor.execute("INSERT INTO ragreedtos (id) VALUES (%s)", (userid,))
+				m = "<@" + userid + "> 利用規約への同意を確認しました。"
+				await client.send_message(message.channel, m)
+				cursor.execute('SELECT * FROM ragreedtos')
+				agreetos = cursor.fetchall()
+				agreetos = str(agreetos)
+				print(agreetos)
+				await client.delete_message(message)
 
 
 			# 保存を実行（忘れると保存されないので注意）
